fixing issue with dead-lock between ray cluster workers/head and kubedirector
kubedirector launches the startscript of each pod one by one, expecting the last to complete
before starting the next one. If the head node is not started first, ray worker wait for the
head node to start and the startscript does not end, thus the head node is not started.

fix is to run a subscript in background from the start script for kudirector to start them
all in parallel and not stay blocked